app/requirements.txt
django
djangorestframework
googlemaps
django-collectionfield